fix: missing tile ids not deleting and state being carried over between request calls?

+++ coral/utils/merge_resources.py
@@ -278,6 +278,9 @@ class MergeResources:
     def merge_resources(
         self, base_resource_id, merge_resource_id, merge_tracker_resource_id, overwrite_multiple_tiles=False
     ):
+        # FIXME: HOW DO REQUESTS MAINTAIN STATE, move to __init__
+        self.merge_map = {}
+
         if not base_resource_id or not merge_resource_id:
             raise "Missing base or merge resource ID"
         
@@ -361,12 +364,12 @@ class MergeResources:
                     base_tile.data = merged_tile_data
                     base_tile.save()
                     continue
-
             # Create the additional tiles for the base resource
             if merge_data["cardinality"] == "n":
                 if overwrite_multiple_tiles:
                     for tile in merge_data["base_tiles"]:
-                        tile.delete()
+                        if tile.tileid:
+                            tile.delete()
                 for tile in merge_data["merge_tiles"]:
                     parent_tile = self.discover_parent_tile(tile)
                     new_tile = Tile(
